Improved Speech Rec + Redundent Code Cleanup
Made several functions: DisableTimers(), EnableTimers() which
enables/disables all timers.

SpeakString() is for quickly making l-e talk. Takes a message parameter
and optionally parameters for rate, volume, and a boolean to stop the
timers while the speech is called and then restart them.

I also created a SendCommand function that takes a string parameter that
will open the serial port, send the command and close it. This function
also has an optional parameter for resetting the timers.